everything runs up until process so need to check this next

+++ tests/tests.sh
@@ -10,34 +10,39 @@ source /root/micromamba/etc/profile.d/micromamba.sh
 micromamba activate locpix-points
 
 # Test preprocess
-if pytest tests/preprocess.py
+if 0
 then
-    echo "preprocess successful"
+    if pytest tests/preprocess.py
+    then
+        echo "preprocess successful"
+    else
+        echo "removing output"
+        rm -r tests/output
+        exit
+    fi
+
+    # Activate correct environment
+    micromamba activate rapids=23.10
+
+    # Test feat extract
+    if pytest tests/featextract.py
+    then
+        echo "feat extract successful"
+    else
+        echo "removing output"
+        rm -r tests/output
+        exit
+    fi
 else
-    echo "removing output"
-    rm -r tests/output
-    exit
+
+    # Activate correct environment
+    micromamba activate locpix-points
+
+    # Test preprocess
+    pytest tests/process.py
+
+    # remove files regardless of last script success
+    #echo "removing output end"
+    #rm -r tests/output
+
 fi
-
-# Activate correct environment
-micromamba activate rapids=23.10
-
-# Test feat extract
-if pytest tests/featextract.py
-then
-    echo "feat extract successful"
-else
-    echo "removing output"
-    rm -r tests/output
-    exit
-fi
-
-# Activate correct environment
-micromamba activate locpix-points
-
-# Test preprocess
-pytest tests/process.py
-
-# remove files regardless of last script success
-#echo "removing output end"
-#rm -r tests/output
